// ui/src/lib/types/AutotileKind.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { AutotileRuleSet } from "./AutotileRuleSet";
import type { PeeringSet } from "./PeeringSet";

/**
 * Autotile algorithm variant for a tileset.
 */
export type AutotileKind = { "kind": "blob47" } | { "kind": "corner16" } | { "kind": "minimal4" } | { "kind": "custom" } & AutotileRuleSet | { "kind": "peering" } & PeeringSet;